jhon

+++ src/app/dashboard/guides/new/page.tsx
@@ -115,7 +115,7 @@ const CREATE_SALE_MUTATION = gql`
         $guideReasonTransfer: String!
         $productIdSet: [Int!]!
         $descriptionSet: [String!]!
-        $quantitySet: [Int!]!
+        $quantitySet: [Float!]!
         $relatedDocumentsSerialSet: [String!]!
         $relatedDocumentsDocumentTypeSet: [String!]!
         $relatedDocumentsCorrelativeSet: [Int!]!
@@ -719,7 +719,7 @@ function NewGuidePage() {
                     (item: any) => item.description || ""
                 ),
                 quantitySet: guide.operationdetailSet.map((item: any) =>
-                    parseInt(item.quantity)
+                    parseFloat(item.quantity)
                 ),
 
                 relatedDocumentsSerialSet: guide.relatedDocuments.map(
